agent/pyproject.toml
[project]
name = "content-filter-agent"
version = "1.0.0"
description = "LiveKit agent for content filtering with AI vision"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "livekit>=0.11.0",
    "livekit-agents[anthropic]>=1.2",
    "openai>=1.0.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310"]
